requirements.txt
absl-py==2.1.0
accelerate==1.0.1
antlr4-python3-runtime==4.9.3
anykeystore==0.2
apex==0.9.10.dev0
Brotli @ file:///croot/brotli-split_1714483155106/work
calflops==0.3.2
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
contourpy==1.3.0
cryptacular==1.6.2
cycler==0.12.1
DateTime==5.5
deeplake==4.0.0
defusedxml==0.7.1
einops==0.8.0
filelock @ file:///croot/filelock_1700591183607/work
fonttools==4.54.1
fsspec==2024.10.0
ftfy==6.3.0
fvcore==0.1.5.post20221221
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
greenlet==3.1.1
grpcio==1.67.0
huggingface-hub==0.26.1
hupper==1.12.1
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.10
Jinja2 @ file:///croot/jinja2_1716993405101/work
joblib==1.4.2
kiwisolver==1.4.7
lightning-utilities==0.11.8
Markdown==3.7
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.9.2
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1717597493534/work
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
packaging==24.1
PasteDeploy==3.1.0
pbkdf2==1.3
peft==0.13.2
pillow @ file:///croot/pillow_1721059439630/work
plaster==1.1.2
plaster-pastedeploy==1.0.1
portalocker==2.10.1
protobuf==5.28.3
psutil==6.1.0
pycocoevalcap==1.2
pycocotools==2.0.8
pyparsing==3.2.0
pyramid==2.0.2
pyramid-mailer==0.15.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
regex==2024.9.11
repoze.sendmail==4.4.1
requests @ file:///croot/requests_1721410876868/work
requests-oauthlib==2.0.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
sentencepiece==0.2.0
six==1.16.0
SQLAlchemy==2.0.36
sympy @ file:///croot/sympy_1724938189289/work
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.5.0
timm==1.0.11
tokenizers==0.20.1
torch==2.3.0
torchaudio==2.3.1
torchmetrics==1.5.1
torchvision==0.18.1
tqdm==4.66.5
transaction==5.0
transformers==4.46.0
translationstring==1.4
triton==2.3.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
urllib3 @ file:///croot/urllib3_1727769808118/work
velruse==1.1.1
venusian==3.1.0
wcwidth==0.2.13
WebOb==1.8.9
Werkzeug==3.0.5
WTForms==3.2.1
wtforms-recaptcha==0.3.2
xformers==0.0.26.post1
yacs==0.1.8
zipp==3.20.2
zope.deprecation==5.0
zope.interface==7.1.1
zope.sqlalchemy==3.1
